+++ NEWS.md
@@ -11,7 +11,7 @@
 * [#98](https://github.com/boxuancui/DataExplorer/issues/98): Added band customization to `plot_missing`.
 * [#100](https://github.com/boxuancui/DataExplorer/issues/100): Switched `geom_text` to `geom_label`.
 * [#103](https://github.com/boxuancui/DataExplorer/issues/103): Report title can now be customized in `create_report`.
-* [#108](https://github.com/boxuancui/DataExplorer/issues/108): Added option to treat binary features as discrete in `plot_bar`.
+* [#108](https://github.com/boxuancui/DataExplorer/issues/108): Added option to treat binary features as discrete in `plot_bar`, `plot_histogram` and `plot_density`.
 
 # DataExplorer 0.7.1
 ## Enhancements